WIP: query 1 usando colas
