+++ system/filter_mapper/filter_mapper.go
@@ -1,193 +1,55 @@
-package filter_mapper
+package main
 
 import (
+	"bytes"
 	"fmt"
 	"malasian_coffe/packet"
-	"math"
-	"strconv"
-	"strings"
-	"time"
+	filter_mapper "malasian_coffe/system/filter_mapper/src"
+
+	amqp "github.com/rabbitmq/amqp091-go"
 )
 
-type filterMapperOptions func(string) string
-
-func filterByYearCommon(input []string) string {
-	final := ""
-	for _, line := range input {
-		data := strings.Split(line, ",")
-		if len(data) < 9 {
-			panic("Invalid data format")
-		}
-		if yearCondition(data) {
-			final += line + "\n"
-		}
+// Argumentos que recibe:
+// 1. Address de rabbit
+// 2. Nombre de la funcion que tiene que ejecutar
+func main() {
+	type Colas struct {
+		colaInput  []string
+		colaOutput []string
 	}
-	return final
-}
+	funcionesYColas := make(map[string]Colas)
+	funcionesYColas["query1YearAndAmount"] = Colas{colaInput: []string{"entrada-1"}, colaOutput: []string{"salida-1"}}
+	// funcionesYColas["query2"] = Colas{colaInput: []string{"entrada-2"}, colaOutput: []string{"salida-2"}}
 
-func yearCondition(data []string) bool {
-	layout := "2006-01-02 15:04:05" // Go's reference layout
-	t, _ := time.Parse(layout, data[8])
-	return t.Year() >= 2024 && t.Year() <= 2025
-}
+	// nombre_funcion := "query1YearAndAmount"
+	//map con key de nombre de la funcion y clave tupla de cola input y cola output
 
-func filterFunctionQuery1(input string) string {
-	lines := strings.Split(input, "\n")
-
-	final := ""
-	for _, line := range lines {
-		if strings.TrimSpace(line) == "" {
-			continue
-		}
-		data := strings.Split(line, ",")
-		if len(data) < 9 {
-			panic("Invalid data format")
-		}
-		amount, _ := strconv.ParseFloat(data[7], 64)
-		amount = math.Round(amount*10) / 10
-		if yearCondition(data) && amount >= 15.0 {
-			final += data[0] + "," + strconv.FormatFloat(amount, 'f', 1, 64) + "\n"
-		}
-	}
-	return final
-}
-
-func filterFunctionQuery2a(input string) string {
-	lines := strings.Split(input, "\n")
-	final := ""
-	for _, line := range lines {
-		data := strings.Split(line, ",")
-		if len(data) < 6 {
-			panic("Invalid data format")
-		}
-		layout := "2006-01-02 15:04:05" // Go's reference layout
-		t, _ := time.Parse(layout, data[5])
-		if t.Year() >= 2024 && t.Year() <= 2025 {
-			final += data[1] + "," + data[2] + "," + data[5] + "\n"
-		}
-	}
-	return final
-}
-
-func filterFunctionQuery2b(input string) string {
-	lines := strings.Split(input, "\n")
-	final := ""
-	for _, line := range lines {
-		data := strings.Split(line, ",")
-		if len(data) < 6 {
-			panic("Invalid data format")
-		}
-		layout := "2006-01-02 15:04:05" // Go's reference layout
-		t, _ := time.Parse(layout, data[5])
-		if t.Year() >= 2024 && t.Year() <= 2025 {
-			final += data[1] + "," + data[4] + "," + data[5] + "\n"
-		}
-	}
-	return final
-}
-
-func mapStoreIdAndName(input string) string {
-	lines := strings.Split(input, "\n")
-	final := ""
-	for _, line := range lines {
-		data := strings.Split(line, ",")
-		if len(data) < 8 {
-			panic("Invalid data format")
-		}
-		final += data[0] + "," + data[1] + "\n"
-	}
-	return final
-}
-
-func filterFunctionQuery3Transactions(input string) string {
-	lines := strings.Split(input, "\n")
-	final := ""
-	for _, line := range lines {
-		if strings.TrimSpace(line) == "" {
-			continue
-		}
-		data := strings.Split(line, ",")
-		if len(data) < 9 {
-			panic("Invalid data format")
-		}
-		layout := "2006-01-02 15:04:05" // Go's reference layout
-		t, _ := time.Parse(layout, data[8])
-		amount, _ := strconv.ParseFloat(data[7], 64)
-		amount = math.Round(amount*10) / 10
-		if yearCondition(data) && t.Hour() >= 6 && t.Hour() <= 23 {
-			final += data[1] + "," + strconv.FormatFloat(amount, 'f', 1, 64) + "," + data[8] + "\n"
-		}
-	}
-	return final
-}
-
-func filterFunctionQuery4Transactions(input string) string {
-	lines := strings.Split(input, "\n")
-	final := ""
-	for _, line := range lines {
-		if strings.TrimSpace(line) == "" {
-			continue
-		}
-		data := strings.Split(line, ",")
-		if len(data) < 9 {
-			panic("Invalid data format")
-		}
-		if yearCondition(data) {
-			final += data[0] + "," + data[1] + "," + data[4] + "\n"
-		}
-	}
-	return final
-}
-
-func filterFunctionQuery4UsersBirthdates(input string) string {
-	lines := strings.Split(input, "\n")
-	final := ""
-	for _, line := range lines {
-		if strings.TrimSpace(line) == "" {
-			continue
-		}
-		data := strings.Split(line, ",")
-		if len(data) < 4 {
-			panic("Invalid data format")
-		}
-		final += data[0] + "," + data[2] + "\n"
-	}
-	return final
-}
-
-type FilterMapper struct {
-}
-
-func (c *FilterMapper) Process(pkt packet.Packet, function string) []packet.Packet {
-	input := pkt.GetPayload()
-	function_name := strings.ToLower(function)
-
-	input_lines := strings.Split(input, "\n")
-
-	var output string
-	switch function_name {
-	case "yearfilter": //Only testing purposes
-		output = filterByYearCommon(input_lines)
-	case "query1yearandamount":
-		output = filterFunctionQuery1(input)
-	case "query2ayearandquantity":
-		output = filterFunctionQuery2a(input)
-	case "query2byearandsubtotal":
-		output = filterFunctionQuery2b(input)
-	case "query3mapstoreidandname":
-		output = mapStoreIdAndName(input)
-	case "query3transactions":
-		output = filterFunctionQuery3Transactions(input)
-	case "query4transactions":
-		output = filterFunctionQuery4Transactions(input)
-	case "query4usersbirthdates":
-		output = filterFunctionQuery4UsersBirthdates(input)
-	default:
-		panic(fmt.Sprintf("Unknown function %s", function))
+	rconn, _ := amqp.Dial("amqp://guest:guest@localhost:5672/")
+	ch, _ := rconn.Channel()
+	ch.QueueDeclare(
+		"entrada-1", // name
+		false,       // durable
+		false,       // delete when unused
+		false,       // exclusive
+		false,       // no-wait
+		nil,         // arguments
+	)
+	msgs, _ := ch.Consume(
+		"entrada-1", // queue
+		"",          // consumer
+		false,       // auto-ack
+		false,       // exclusive
+		false,       // no-local
+		false,       // no-wait
+		nil,         // args
+	)
+	worker := filter_mapper.FilterMapper{}
+	for message := range msgs {
+		packet_reader := bytes.NewReader(message.Body)
+		packet, _ := packet.DeserializePackage(packet_reader)
+		fmt.Printf("HOLA ESTAS EN EL FILTER MAPPER %v\n", packet)
+		result := worker.Process(packet, "query1YearAndAmount")
+		fmt.Printf("HOLA YA PROCESASTE EL PAQUETE %v\n", result)
 	}
 
-	outputs := []string{output}
-	new_packets := packet.ChangePayload(pkt, outputs)
-
-	return new_packets
 }
