+++ system/system.go
@@ -2,13 +2,8 @@ package main
 
 import (
 	"fmt"
-	"bytes"
-	"net"
-	"os"
 
 	"malasian_coffe/packet"
-
-	"malasian_coffe/utils/network"
 
 	amqp "github.com/rabbitmq/amqp091-go"
 )
@@ -29,23 +24,38 @@ func main() {
 		panic(fmt.Errorf("Failed to connect to the channel: %w", err))
 	}
 	ch.QueueDeclare(
-		"prueba",  // name
-		false, // durable
-		false, // delete when unused
-		false, // exclusive
-		false, // no-wait
-		nil,   // arguments
+		"entrada-1", // name
+		false,       // durable
+		false,       // delete when unused
+		false,       // exclusive
+		false,       // no-wait
+		nil,         // arguments
 	)
 
-	rabbit_addr := os.Args[1]
+	/*	rabbit_addr := os.Args[1]
 
-	list, err := net.Listen("tcp", rabbit_addr)
-	if err != nil {
-		panic(fmt.Sprintf("Failed to create listener. Error: %s", err))
-	}
-	conn, _ := list.Accept()
+		list, err := net.Listen("tcp", rabbit_addr)
+		if err != nil {
+			panic(fmt.Sprintf("Failed to create listener. Error: %s", err))
+		}
+		conn, _ := list.Accept()*/
 
-	for {
+	transactionsRaw := []string{"2ae6d188-76c2-4095-b861-ab97d3cd9312,4,5,,,38.0,0.0,38.0,2024-07-01 07:00:00\n" +
+		"7d0a474d-62f4-442a-96b6-a5df2bda8832,7,1,,,33.0,0.0,33.0,2025-07-01 07:00:02\n" +
+		"928498fd-edbf-456c-bbd5-31aa56dc96c9,8,1,,,14.0,0.0,14.0,2023-07-01 07:02:21\n" +
+		"48968d91-dd5a-47f2-8646-42f8b587932f,3,1,,,30.0,0.0,30.0,2023-07-01 07:01:54"}
+	pkt_empty := packet.Packet{}
+	pkt := packet.ChangePayload(pkt_empty, transactionsRaw)[0]
+
+	fmt.Printf("%v\n", pkt)
+	ch.Publish("", "entrada-1", false, false,
+		amqp.Publishing{
+			// DeliveryMode: amqp.Persistent,
+			ContentType: "text/plain",
+			Body:        pkt.Serialize(),
+		})
+
+	/*for {
 		packet_b, err := network.ReceiveFromNetwork(conn)
 		packet_reader := bytes.NewReader(packet_b)
 		if err != nil {
@@ -62,5 +72,5 @@ func main() {
 			ContentType: "text/plain",
 			Body:        packet.Serialize(),
 		})
-	}
+	}*/
 }
